feat: Show open folder button in success alert after move or duplicate
